pyproject.toml
[tool.poetry]
name = "clashcode"
version = "0.1.0"
description = "AI编码原生安全助手 - 红蓝对抗代码审查 + 全局影响分析"
authors = ["ClashCode Team"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "clashcode"}]

[tool.poetry.scripts]
clashcode = "clashcode.cli.main:app"
clashcode-mcp = "clashcode.mcp.server:main"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.0.0"
langgraph = ">=0.2.0"
langchain-core = ">=0.3.0"
langchain-openai = ">=0.2.0"
langchain-anthropic = ">=0.2.0"
tree-sitter = ">=0.21.0"
tree-sitter-python = ">=0.21.0"
pyyaml = "^6.0"
pydantic = "^2.0"
mcp = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 120
